requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
anthropic==0.18.1
python-dotenv==1.0.0